Minor Update
Added Movement Pad, still in early stages same with interact button. Still fixing on how to bind movement to the movement pad. Temporarily the player can move by dragging the left side of the screen and rotating the camera using the right side. Also fixed the UI for Menu, Select Level, and Terms and Conditions. About screen and Instructions screen also needs to be fixed. Will continue to update ASAP.